+++ src/com/inet/gradle/setup/unix/rpm/RpmBuilder.java
@@ -323,6 +323,9 @@ public class RpmBuilder extends UnixBuilder<Rpm, SetupBuilder> {
             if( starter.getCategories() != null ) {
                 fw.write( "Categories=" + starter.getCategories() + "\n" );
             }
+            for( String additionalSetting : starter.getAdditionalSettings() ) {
+                fw.write( additionalSetting + "\n" );
+            }
         }
 
         // register the mime type and the default app for the extensions